Add AI review functionality and context file handling
- Implemented generateAIReview and processAIReviewComments functions for AI-driven PR reviews.
- Added getContextFilesContent function to retrieve static files for PR context.
- Created unit tests for GitHub client initialization and options.
- Refactored main action logic to utilize new helper functions for better modularity.
- Updated TypeScript configuration to include noEmit and allowImportingTsExtensions options.
